Update streak logic
Problem statement: we took a day for maintenance, and all player streaks were broken due to hardcoded date-based logic, which forced a correct answer on the previous calendar day to maintain streak.

Fix: this date-based logic is removed and broken streaks are reset at the end of day for all players who did not answer that day's question correctly.

Effect: the db is always correct after play has concluded.